Server/ImageProcessing/requirements.txt
matplotlib==3.0.2
opencv_python==4.2.0.34
scikit_image==0.16.2
Pillow==5.3.0
